+++ feature/fido2/src/androidMain/kotlin/com/chimali/fido2/data/transport/Fido2Transport.kt
@@ -8,6 +8,8 @@ interface Fido2Transport {
 
     suspend fun connect(): Result<Unit>
 
+    suspend fun connectDevice(macAddress: String): Result<Unit>
+
     suspend fun disconnect(): Result<Unit>
 
     suspend fun sendCommand(command: ByteArray): Result<ByteArray>